+++ mopidy/core/__init__.py
@@ -2,8 +2,8 @@ from __future__ import unicode_literals
 
 # flake8: noqa
 from .actor import Core
-from .current_playlist import CurrentPlaylistController
 from .library import LibraryController
 from .listener import CoreListener
 from .playback import PlaybackController, PlaybackState
 from .stored_playlists import StoredPlaylistsController
+from .tracklist import TracklistController